image-search/requirements.txt
open_clip_torch>=2.24.0
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.0.0
requests>=2.31.0
duckdb>=0.9.2
numpy>=1.26.0
tqdm>=4.66.0
streamlit>=1.30.0
